+++ packages/file-bundle/config/file-zip.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
  */
 
 use Rekalogika\Contracts\File\FileRepositoryInterface;
+use Rekalogika\File\Zip\DirectoryResourceServer;
 use Rekalogika\File\Zip\FileZip;
 use Rekalogika\File\Zip\ZipDirectory;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
@@ -25,11 +26,24 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $services->set(FileZip::class)
         ->args([
             service(ZipDirectory::class),
+            service(TranslatorInterface::class)
         ]);
 
     $services->set(ZipDirectory::class)
         ->args([
             service(FileRepositoryInterface::class),
             service(TranslatorInterface::class)->nullOnInvalid(),
+        ]);
+
+    $services->set(DirectoryResourceServer::class)
+        ->args([
+            service(FileZip::class),
+            service(TranslatorInterface::class),
+        ])
+        ->tag('rekalogika.temporary_url.resource_server', [
+            'method' => 'respond',
+        ])
+        ->tag('rekalogika.temporary_url.resource_transformer', [
+            'method' => 'transform',
         ]);
 };