+++ packages/file-contracts/src/FilesInterface.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace Rekalogika\Contracts\File;
 
+use Symfony\Contracts\Translation\TranslatableInterface;
+
 /**
  * Represents a container containing instances of FileInterface
  *
@@ -22,4 +24,9 @@ interface FilesInterface extends
     \Traversable,
     \Countable
 {
+    /**
+     * The name for this set of files, will be used as the name of the
+     * downloaded folder or archive file, etc.
+     */
+    public function getName(): \Stringable&TranslatableInterface;
 }